refactor(polymarket): extract Gamma client into a separate module
Move `fetchGammaApi`, `fetchMarketsFromApi`, `fetchEventFromApi`, and
the `PolymarketMarket` / `PolymarketEvent` / `PolymarketSeries` /
`GammaEvent` interfaces from `services/polymarket/index.ts` into a new
`services/polymarket/gamma.ts`. The scraper loop in `index.ts` re-imports
them from there.

Fixes the CI test failure introduced by `fetch-gamma.test.ts`: importing
`./index` at the top level of that test file pre-loaded `lib/clickhouse`
and `lib/batch-insert` with their real implementations, so when
`index.test.ts` later registered partial `mock.module` replacements,
bun's subsequent re-evaluation of `./index` failed to resolve the
stripped exports. The new test imports `./gamma` directly, which has no
ClickHouse / batch-insert dependencies, so module loading order no
longer matters.
